Add files via upload
Receipts.US.pdf is a handwritten overview of how the project will be started. The plan is not very technical and goes over the general idea of how the app will be made. The technical details for each aspect must be reviewed in order to be planned accurately and constructed correctly.